backend/requirements.txt
fastapi
uvicorn
qdrant-client
sentence-transformers
torch
openai
python-dotenv
pydantic
python-multipart
PyPDF2
python-docx
requests
